ST-126-refactor: 데이터 페칭을 zustand에서함
