+++ src/store/socialLoginStore.ts
@@ -2,7 +2,6 @@ import Cookies from "js-cookie";
 import { create } from "zustand";
 import { persist } from "zustand/middleware";
 
-// 상태 타입 정의
 export interface UserInfo {
   id: string | null;
   email: string | null;
@@ -10,34 +9,41 @@ export interface UserInfo {
   profileImageUrl: string | null;
   createdAt: string | null;
   updatedAt: string | null;
+  social: boolean;
 }
 
 interface SocialLoginStoreType extends UserInfo {
-  social: boolean;
   socialLogin: (userInfo: UserInfo) => void;
   commonLogin: (userInfo: UserInfo) => void;
+  logout: () => void;
 }
+
+const initialState: UserInfo = {
+  id: null,
+  email: null,
+  nickname: null,
+  profileImageUrl: null,
+  createdAt: null,
+  updatedAt: null,
+  social: false,
+};
 
 const socialLoginStore = create<SocialLoginStoreType>()(
   persist(
     set => ({
-      social: false,
-      id: null,
-      email: null,
-      nickname: null,
-      profileImageUrl: null,
-      createdAt: null,
-      updatedAt: null,
+      ...initialState,
       socialLogin: (userInfo: UserInfo) =>
         set({
-          social: true,
           ...userInfo,
+          social: true,
         }),
+
       commonLogin: (userInfo: UserInfo) =>
         set({
-          social: false,
           ...userInfo,
+          social: false,
         }),
+      logout: () => set(initialState),
     }),
     {
       name: "social-login-store",
